refactor by swapping execute- and do- move
